fix(trusty-controller): remediate REQUEST_CHANGES review findings on PR #1075
1. LICENSE contradiction: fix Cargo.toml to use `license.workspace = true`
   (MIT, per #898 relicense + root Cargo.toml + all sibling crates). Remove
   the incorrect Elastic-2.0 ADR-0006 comment. Fix README.md License section
   to say MIT. LICENSE file was already MIT so no change needed there.

2. --json-port vs --json precedence: document and enforce at runtime in
   `commands::port::run`. `--json-port` takes precedence over global `--json`
   when both are present; clap's `conflicts_with` cannot enforce this across
   subcommand boundaries for `global = true` flags, so precedence is applied
   in the handler and documented in both `cli.rs` Port variant doc and
   `commands/port.rs` module header. New tests assert both parse orderings
   succeed and that `json_port` wins in the handler.

3. Unused NotYetImplemented struct: wire it into `print_not_yet_implemented`
   as the canonical JSON serialization type for all stub output. The function
   now constructs a `NotYetImplemented` value and serialises it, removing the
   ad-hoc `serde_json::json!({})` inline object. Both paths (human + JSON)
   now go through the typed struct.

4. scope detection unused: add `// TODO(#920 / Phase 1):` comment and
   `let _ = cli.scope` suppression in `main.rs` dispatch, explaining that
   scope wiring lands when the manifest-driven probe loop is implemented.

Also split the inline test module from `cli.rs` into `cli_tests.rs` to keep
`cli.rs` under the 500-line file cap (line-cap gate now passes: 0 violations).
